pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["Sergey Klimov", "Maxim Rukhlov", "Alex Erdyakov"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
psycopg2-binary = "^2.9.1"
sqlalchemy = "1.4.22" # temp fix
Flask-SQLAlchemy = "^2.5.1"
flask-restx = "^0.5.0"
python-decouple = "^3.4"
Flask-Bcrypt = "^0.7.1"
Flask-Migrate = "^3.1.0"
PyJWT = "^2.1.0"
factory-boy = "^3.2.0"
gevent = "^21.8.0"
uWSGI = "^2.0.19"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
isort = "^5.9.3"
flake8 = "^3.9.2"
pytest-cov = "^2.12.1"
ipython = "^7.26.0"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
